2,DjangoAPI/requirements.txt
asgiref==3.9.2
attrs==25.4.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0